03-multiagent-demo/requirements.txt
strands-agents>=1.27.0
strands-agents[openai]>=1.27.0
